Revert "fix: resolve CodeQL go/incorrect-integer-conversion (#229)"
This reverts commit f53a33c1b4cd008b6658f2f5c385e22ba6255366.

+++ internal/process/hosts.go
@@ -110,7 +110,7 @@ func createHostSamples(config *config.Config) {
 			memoryUsed := host.Summary.QuickStats.OverallMemoryUsage
 			checkError(config.Logrus, ms.SetMetric("mem.usage", memoryUsed, metric.GAUGE))
 
-			memoryFree := memoryTotal - int64(memoryUsed)
+			memoryFree := int32(memoryTotal) - memoryUsed
 			checkError(config.Logrus, ms.SetMetric("mem.free", memoryFree, metric.GAUGE))
 
 			// cpu
